Use JSON logging by default
This change adds the option to choose which logging format you want to
use. The former unstructured log is still available but the new default
is a structured log in the form of JSON.

Exceptions still behave the same as before. Maybe some day they will
also be in JSON format.

Additionally, there is now a dedicated log line for every incoming
HTTP request.

The log line mentioning that a file will be scanned was set to debug
level to reduce the amount of useless logs in day2day opertions.